+++ Content.Client/IoC/ClientContentIoC.cs
@@ -20,7 +20,8 @@ using Content.Shared.Module;
 
 using Content.Client.White.JoinQueue;
 using Content.Client.White.Sponsors;
-using Content.Client.White.Trail;
+using Content.Client.White.Trail.Line;
+using Content.Shared.White.Line;
 
 namespace Content.Client.IoC
 {
@@ -49,7 +50,7 @@ namespace Content.Client.IoC
             IoCManager.Register<SponsorsManager>();
             IoCManager.Register<JoinQueueManager>();
 
-            IoCManager.Register<ITrailLineManager<ITrailLine>, TrailLineManager<TrailLine>>(); //prikol
+            IoCManager.Register<ITrailLineManager<ITrailLine>, TrailLineManager<TrailLineStretch>>(); //prikol
         }
     }
 }
